backend/requirements.txt
flask
flask-cors
flask-jwt-extended
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pymongo
bcrypt
python-dotenv
google-genai
scikit-learn
pandas
numpy
mongomock
PyPDF2
PyJWT
google-auth
google-auth-oauthlib
google-api-python-client
